Updated sql html

+++ src/main/java/com/company_mngm_sys/demo/dao/EmployeeRepo.java
@@ -9,6 +9,6 @@ public interface EmployeeRepo extends CrudRepository<Employees, Long> {
     @Override
     public List<Employees> findAll();
 
-    public Employees findByEmployeeId(Integer id);
+    public Employees findByEmployeeId(Long id);
 
 }
